PLUS-stable: Merged 1777-1805,1807-1810,1812-1846 from plus-trunk into +STABLE+
........
  r1809 | tdsoul | 2007-10-03 19:52:45 -0400 (Wed, 03 Oct 2007) | 9 lines
  
  PLUS-trunk:
   * SITE RESCAN gets some new options:
     * -r[ecursive] : Recursive mode extends scan into subdirectories.
     * -q[uiet]     : Quiet output, only display missing/bad files.
     * -c[ached]    : Use only cached checksums (lightning fast).
     * -delete      : Delete bad files you have permission to delete.
  
  Thanks to Dom for allowing me to use his implentation as a reference.
........
  r1810 | tdsoul | 2007-10-03 20:45:38 -0400 (Wed, 03 Oct 2007) | 4 lines
  
  PLUS-trunk:
   * Add -missingdir option to tommie`'s FindNuke mod.  This new option makes it easier to find dirs which are missing a required subdirectory.
   * Fix help info I previously failed to add when submitting tommie`s original mod.
........
  r1820 | tdsoul | 2007-10-22 19:05:29 -0400 (Mon, 22 Oct 2007) | 10 lines
  
  PLUS-trunk:
   * Deny reqdel on a non-empty directory, decrement request counter when successful.
     * TODO: Decrement the request counter of the user who created the request, not the user who issued the command [only a problem when user.isAdmin() is true]
   * Fixed a possible NPE in replaceIdent() created for r1806
   * Fixed a logic bug in IRCUsermanagement finduser command.
   * Misc. non-functional changes. Nothing spectacular, mostly formatting / style edits which will make it easier to share code between my public/private repositories.
  
  Thanks to _Dink_ and N0S on IRC for the nudge.
  
  Closes: #173
........
  r1821 | tdsoul | 2007-11-01 18:40:18 -0400 (Thu, 01 Nov 2007) | 7 lines
  
  PLUS-trunk: 
   * REQDEL (See #179)
     * Update error message when failing to delete a dir because it is empty, clean up some of the logic to make it easier to read.
     * While I'm here anyway, fix it so that it credits request count for user who owns the request, instead of the user who initiated REQDEL command (in the case of an admin)
........
  r1824 | tdsoul | 2007-11-03 00:26:55 -0400 (Sat, 03 Nov 2007) | 4 lines
  
  PLUS-trunk: 
   * REQDEL (See #179)
     * Fix a logic bug, and finish editing the notempty response, apparently I prematurely comitted last time :D
........

Also a previous attempt to merge r1811 failed to add a new file, added here.



git-svn-id: https://drftpd.svn.sourceforge.net/svnroot/drftpd/plus/2.0/branches/+STABLE+@1847 c83153ee-f3e8-0310-b165-dc3a72e09794

+++ src/net/drmods/plugins/irc/Request.java
@@ -36,6 +36,7 @@ import org.drftpd.dynamicdata.Key;
 import org.drftpd.plugins.SiteBot;
 import org.drftpd.remotefile.LinkedRemoteFileInterface;
 import org.drftpd.sitebot.IRCCommand;
+import org.drftpd.usermanager.NoSuchUserException;
 import org.drftpd.usermanager.User;
 import org.drftpd.usermanager.UserFileException;
 import org.tanesha.replacer.ReplacerEnvironment;
@@ -259,7 +260,7 @@ public class Request extends IRCCommand {
             dir.createDirectory("REQUEST-by." + requser + "-" + dirName);
             LinkedRemoteFileInterface reqdir = dir.getFile("REQUEST-by." + requser + "-" + dirName);
             reqdir.setOwner(requser);
-            user.getKeyedMap().setObject(Request.REQUESTS, user.getKeyedMap().getObjectInt(Request.REQUESTS)+1);;
+            user.getKeyedMap().setObject(Request.REQUESTS, user.getKeyedMap().getObjectInt(Request.REQUESTS)+1);
             ArrayList<String> forceToChannels = new ArrayList<String>();
             forceToChannels.add(msgc.getDest());
             getGlobalContext().dispatchFtpEvent(new DirectorySiteBotEvent(
@@ -315,23 +316,38 @@ public class Request extends IRCCommand {
             
             for (Iterator iter = dirs.iterator(); iter.hasNext();) {
                 LinkedRemoteFileInterface file = (LinkedRemoteFileInterface) iter.next();
-                if (file.isDirectory()) {
-                    if (file.getName().endsWith(dirName)) {
-                        nodir = false;
-                        if (file.getUsername().equals(user.getName()) || user.isAdmin()) {
-                            file.delete();
-                            deldir = true;
-                            ArrayList<String> forceToChannels = new ArrayList<String>();
-                            forceToChannels.add(msgc.getDest());
-                            getGlobalContext().dispatchFtpEvent(new DirectorySiteBotEvent(   
-                            		user, "REQDEL", file, forceToChannels));   
-                            break;
-                        } else {
-                            out.add(ReplacerUtils.jprintf("reqdel.notowner", env, Request.class));
-                            break;
-                        }
-                    } else nodir = true;
-                }
+                if (file.isDirectory() && file.getName().endsWith(dirName)) {
+                    nodir = false;
+                    if (file.dirSize() == 0) {
+                    	if (user.isAdmin() || file.getUsername().equals(user.getName())) {
+	                    	User requestowner = null;
+								try {
+									requestowner = getGlobalContext().getUserManager().getUserByName(file.getUsername());
+								} catch (NoSuchUserException e) {
+								} catch (UserFileException e) {
+								}
+	                        file.delete();
+	                        deldir = true;
+	                        ArrayList<String> forceToChannels = new ArrayList<String>();
+	                        forceToChannels.add(msgc.getDest());
+	                        getGlobalContext().dispatchFtpEvent(new DirectorySiteBotEvent(
+	                        		user, "REQDEL", file, forceToChannels));
+	                        // decrement request counter
+	                        if (requestowner != null) {
+	                        	int reqsMade = requestowner.getKeyedMap().getObjectInt(Request.REQUESTS);
+	                        	if (reqsMade > 0)
+	                        		requestowner.getKeyedMap().setObject(Request.REQUESTS, --reqsMade);
+	                        }
+	                        break;
+	                    } else {
+	                        out.add(ReplacerUtils.jprintf("reqdel.notowner", env, Request.class));
+	                        break;
+	                    }
+                    } else {
+                        out.add(ReplacerUtils.jprintf("reqdel.notempty", env, Request.class));
+                        break;
+	                }
+                } else nodir = true;
             }
             
             if (nodir && !deldir)
